+++ packages/algob/src/internal/deployer.ts
@@ -31,6 +31,7 @@ import type {
 	LsigInfo,
 	RuntimeEnv,
 	SCParams,
+	TxnReceipt,
 } from "../types";
 import { DeployerConfig } from "./deployer_cfg";
 
@@ -185,27 +186,6 @@ class DeployerBasicMode {
 	 */
 	async loadLogicByFile(name: string, scTmplParams?: SCParams): Promise<LogicSigAccount> {
 		return await getLsig(name, this.algoOp.algodClient, scTmplParams);
-	}
-
-	/**
-	 * Loads logic signature from cache for contract mode. This helps user to avoid
-	 * passing templaste parameters always during loading logic signature.
-	 * @param name ASC name
-	 * @returns loaded logic signature from artifacts/cache/<file_name>.teal.yaml
-	 * @deprecated this function will be removed in the next release. Use mkContractLsig to
-	 * store lsig info in checkpoint (against lsigName), and query it in scripts using
-	 * getLsig
-	 */
-	async loadLogicFromCache(name: string): Promise<LogicSigAccount> {
-		return await getLsigFromCache(name);
-	}
-
-	/**
-	 * Alias to `this.compileASC` with last two parameters being swapped.
-	 * @deprecated this function will be removed in the next release.
-	 */
-	ensureCompiled(name: string, force?: boolean, scTmplParams?: SCParams): Promise<ASCCache> {
-		return this.compileASC(name, scTmplParams, force);
 	}
 
 	/**
@@ -502,7 +482,7 @@ class DeployerBasicMode {
 	 * @param txns list transaction in group
 	 * @returns confirmed tx info of group
 	 */
-	async getReceiptTxns(txns: Transaction[]): Promise<ConfirmedTxInfo[]> {
+	async getReceiptTxns(txns: Transaction[]): Promise<TxnReceipt[]> {
 		return await this.algoOp.getReceiptTxns(txns);
 	}
 }
@@ -890,7 +870,7 @@ export class DeployerDeployMode extends DeployerBasicMode implements Deployer {
 	 */
 	async executeTx(
 		transactions: wtypes.ExecParams[] | wtypes.TransactionAndSign[]
-	): Promise<ConfirmedTxInfo[]> {
+	): Promise<TxnReceipt[]> {
 		return await executeTx(this, transactions);
 	}
 }
@@ -1077,7 +1057,7 @@ export class DeployerRunMode extends DeployerBasicMode implements Deployer {
 	 */
 	async executeTx(
 		transactions: wtypes.ExecParams[] | wtypes.TransactionAndSign[]
-	): Promise<ConfirmedTxInfo[]> {
+	): Promise<TxnReceipt[]> {
 		return await executeTx(this, transactions);
 	}
 }